API Reference

Data Structure

ADSeismic.AcousticPropagator — Type

AcousticPropagator

A structure holding all the information for acoustic simulation

Simulation

ADSeismic.pml_helper — Method

pml_helper(t::Float64, USE_PML_MIN::Bool, USE_PML_MAX::Bool, η::Float64, damping::Float64, NPOWER::Int64)

Computing the PML profile.

I/O

Utilities

ADSeismic.Ricker — Function

Ricker(epp::Union{ElasticPropagatorParams, AcousticPropagatorParams}, 
a::Union{PyObject, <:Real}, 
shift::Union{PyObject, <:Real}, 
amp::Union{PyObject, <:Real}=1.0)

Returns a Ricker wavelet (a tensor).

epp: a ElasticPropagatorParams or an AcousticPropagatorParams

a: Width parameter

shift: Center of the Ricker wavelet

amp: Amplitude of the Ricker wavelet

\[f(x) = \mathrm{amp}A (1 - x^2/a^2) exp(-x^2/2 a^2)\]

where

\[A = 2/sqrt(3a)pi^1/4\]

ADSeismic.compute_loss_and_grads_GPU — Method

compute_loss_and_grads_GPU(model::Function, src::Union{Array{AcousticSource},Array{ElasticSource}},
rcv_sim::Union{Array{AcousticReceiver}, Array{ElasticReceiver}}, Rs::Array{Array{Float64}}, vs::Union{PyObject, Array{PyObject}})

Computes the loss and the gradients of the model on all available GPUs.

model, src, rcv_sim: Model, Source and Receiver returned by load_acoustic_source, load_acoustic_model, load_acoustic_receiver or load_elastic_source, load_elastic_model, load_elastic_receiver

Rs: True receiver data

vs: Variables to be optimized. It or its elements must have ref type.

ADSeismic.sampling_compute_loss_and_grads_GPU_v2 — Method

models : nsrc * batch_size

src : nsrc

rcv_sim : nsrc * batch_size

Rs : nsrc * sample_size

ADSeismic.sampling_compute_loss_and_grads_GPU_v3 — Method

Mathcing statistics.

models : nsrc * batch_size

src : nsrc

rcv_sim : nsrc * batch_size

Rs : nsrc * sample_size

Optimization

ADSeismic.LBFGS! — Method

LBFGS!(sess::PyObject, loss::PyObject, grads::Union{Array{T},Nothing,PyObject}, 
    vars::Union{Array{PyObject},PyObject}; IT_PLOT::Int64 = 10, NX::Int64, NY::Int64, kwargs...) where T<:Union{Nothing, PyObject}

L-BFGS algorithm for optimizing variables.

sess: A working session for current computation

loss: Loss function

grads: gradients $\frac{\partial L}{\partial v}$, where $v$ is the unknown

vars: unknown variables

max_iter: maximum number of iterations

Keyword arguments

IT_PLOT: the plotting frequency.

DIR: the directory for saving the plots. By default, DIR=tmp

NX, NY: the numbers of DOFs in $x$ and $y$ direction

Additional keyword arguments are passed to Optim.optimize